+++ assembly/src/ast/tests.rs
@@ -3,7 +3,7 @@ use alloc::{string::ToString, sync::Arc, vec::Vec};
 use pretty_assertions::assert_eq;
 
 use crate::{
-    Felt, Span, assert_diagnostic, assert_diagnostic_lines,
+    Felt, LibraryNamespace, LibraryPath, Span, assert_diagnostic, assert_diagnostic_lines,
     ast::*,
     diagnostics::{Report, reporting::PrintDiagnostic},
     regex, source_file,
@@ -36,9 +36,9 @@ macro_rules! exec {
     ($name:path) => {{
         let path = stringify!($name);
         let (module, name) = path.split_once("::").expect("invalid procedure path");
-        let name =
-            Ident::new_unchecked(Span::unknown(Arc::from(name.to_string().into_boxed_str())));
-        let name = ProcedureName::new_unchecked(name);
+        let name = Ident::new(Span::unknown(Arc::from(name.to_string().into_boxed_str())))
+            .expect("invalid identifier");
+        let name = ProcedureName::new(name).expect("invalid procedure name");
 
         inst!(Exec(InvocationTarget::ProcedurePath { name, module: module.parse().unwrap() }))
     }};
@@ -53,7 +53,7 @@ macro_rules! call {
     ($name:path) => {{
         let path = stringify!($name);
         let (module, name) = path.split_once("::").expect("invalid procedure path");
-        let name = ProcedureName::new_unchecked(Default::default(), name);
+        let name = ProcedureName::new(Default::default(), name).expect("invalid procedure name");
 
         inst!(Call(InvocationTarget::ProcedurePath { name, module }))
     }};
@@ -1170,4 +1170,126 @@ fn assert_parsing_line_unexpected_token() {
         "  `----",
         r#" help: expected "@", or "begin", or "const", or "export", or "proc", or "use", or end of file, or doc comment"#
     );
+}
+
+/// This test evaluates that we get the expected formatted Miden Assembly output when parsing some
+/// Miden Assembly source code into the AST, and then formatting the AST.
+///
+/// NOTE: Due to current limitations of the parser, round-tripping is currently somewhat lossy:
+///
+/// - Line comments (i.e. not docstrings) are not preserved, and so do not end up in the output
+/// - The original choice to place a sequence of instructions on the same line or multiple lines is
+///   not preserved in the AST, so the formatter always places them on individual lines.
+/// - References to constant values by name are replaced with their value during semantic analysis,
+///   so no named constants appear in the formatted output.
+/// - Constant declarations are not preserved by the parser, and so are not shown in the output
+#[test]
+fn test_roundtrip_formatting() {
+    let source = "\
+#! module doc
+#!
+#! with spaces
+
+#! constant doc
+#!
+#! with spaces
+const.DEFAULT_CONST=100
+
+#! Perform `a + b`, `n` times
+#!
+#! with spaces
+proc.add_n_times # [n, b, a]
+    dup.0
+    push.0
+    u32gt
+    if.true
+        push.0.1
+        while.true  # [total, n, b, a]
+            dup.3 dup.3
+            u32wrapping_add3 # [total', n, b, a]
+            swap.1
+            push.1
+            u32overflowing_sub  # [overflowed, n - 1, total', b, a]
+            swap.1 movdn.3      # [overflowed, total', n', b, a]
+            push.0              # [0, overflowed, total, n', total', b, a]
+            dup.1               # [overflowed, 0, overflowed, total', n', b, a]
+            cdrop               # [continue, total', n', b, a]
+        end
+        movdn.3
+        drop drop drop
+    else
+        u32wrapping_add
+    end
+end
+
+begin
+    push.1.1.DEFAULT_CONST
+    exec.add_n_times
+    push.20
+    assert_eq
+
+    trace.DEFAULT_CONST
+end
+";
+
+    let context = TestContext::default();
+    let source = source_file!(&context, source);
+
+    let module = Module::parse(
+        LibraryPath::new_from_components(LibraryNamespace::Exec, []),
+        ModuleKind::Executable,
+        source,
+    )
+    .unwrap_or_else(|err| panic!("{err}"));
+
+    let formatted = module.to_string();
+    let expected = "\
+#! module doc
+#!
+#! with spaces
+
+#! Perform `a + b`, `n` times
+#!
+#! with spaces
+proc.add_n_times
+    dup.0
+    push.0
+    u32gt
+    if.true
+        push.0
+        push.1
+        while.true
+            dup.3
+            dup.3
+            u32wrapping_add3
+            swap.1
+            push.1
+            u32overflowing_sub
+            swap.1
+            movdn.3
+            push.0
+            dup.1
+            cdrop
+        end
+        movdn.3
+        drop
+        drop
+        drop
+    else
+        u32wrapping_add
+    end
+end
+
+begin
+    push.1
+    push.1
+    push.100
+    exec.add_n_times
+    push.20
+    assert_eq
+    trace.100
+end
+";
+
+    assert_eq!(&formatted, expected);
 }